restircting some pages

+++ base/views.py
@@ -8,6 +8,7 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 
 
@@ -20,6 +21,12 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def login_page(request):
+    
+    # preventing relogin 
+    if request.user.is_authenticated:
+        return redirect ("home")
+    
+    
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -71,6 +78,7 @@ def room(request, pk):
     context = {'room':room}
     return render (request, 'base/room.html', context)
 
+@login_required(login_url = "/login")
 # create_rooom
 def create_room(request):
     form = RoomForm
@@ -83,10 +91,15 @@ def create_room(request):
     context = {"form" : form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url = "/login")
 # function that will update the room
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host: # blocking unauthorized
+        return HttpResponse("You are not the authorized to update")
     
     # updating the room 
     if request.method =="POST":
@@ -98,10 +111,16 @@ def update_room(request, pk):
     context = {"form":form}
     return render(request, "base/room_form.html", context)
 
-# deleting room function
 
+# deleting room function
+@login_required(login_url = "/login")
 def delete_room(request, pk):
     room = Room.objects.get(id = pk)
+    
+    if request.user != room.host: # blocking unauthorized
+      return HttpResponse("You are not the authorized to delete ")
+    
+    
     if request.method == "POST":
         room.delete()
         return redirect ("home")
